Refactoring of event bus into a clearer and cleaner form. Logic has been simplified and teased out. Some inner classes have been factored out into independent classes.
