+++ src/main/asciidoc/dataobjects.adoc
@@ -276,6 +276,10 @@ Sets whether or not the current link is hidden.
 +++
 Sets the argument index.
 +++
+|[[multiValued]]`multiValued`|`Boolean`|
++++
+Sets whether or not the argument can receive several values. Only the last argument can receive several values.
++++
 |[[required]]`required`|`Boolean`|
 +++
 Sets whether or not the current link is required.
@@ -452,6 +456,11 @@ Set the store as a buffer
 +++
 Sets te arg name for this option.
 +++
+|[[choices]]`choices`|`Array of String`|
++++
+Adds a choice to the list of values accepted by this option. If the value set by the user does not match once of these
+ values, a link exception is thrown.
++++
 |[[defaultValue]]`defaultValue`|`String`|
 +++
 Sets the default value of this option
@@ -467,6 +476,10 @@ Configures the current link to be a flag. It will be evaluated to <code>true</co
  <code><pre>
    option.setFlag(true).setSingleValued(true)
  </pre></code>
++++
+|[[help]]`help`|`Boolean`|
++++
+Sets whether or not this option is a "help" option
 +++
 |[[hidden]]`hidden`|`Boolean`|
 +++
@@ -619,25 +632,6 @@ Set whether Netty pooled buffers are enabled
 +++
 |===
 
-[[MetricsOptions]]
-== MetricsOptions
-
-++++
- Vert.x metrics base configuration, this class can be extended by provider implementations to configure
- those specific implementations.
-++++
-'''
-
-[cols=">25%,^25%,50%"]
-[frame="topbot"]
-|===
-^|Name | Type ^| Description
-|[[enabled]]`enabled`|`Boolean`|
-+++
-Set whether metrics will be enabled on the Vert.x instance.
-+++
-|===
-
 [[ClientOptionsBase]]
 == ClientOptionsBase
 
@@ -734,6 +728,25 @@ Set the trust options in jks format, aka Java trustore
 |[[usePooledBuffers]]`usePooledBuffers`|`Boolean`|
 +++
 Set whether Netty pooled buffers are enabled
++++
+|===
+
+[[MetricsOptions]]
+== MetricsOptions
+
+++++
+ Vert.x metrics base configuration, this class can be extended by provider implementations to configure
+ those specific implementations.
+++++
+'''
+
+[cols=">25%,^25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[enabled]]`enabled`|`Boolean`|
++++
+Set whether metrics will be enabled on the Vert.x instance.
 +++
 |===
