+++ src/main/asciidoc/java/streams.adoc
@@ -183,3 +183,4 @@ This has the same meaning as `link:../../apidocs/io/vertx/core/streams/WriteStre
 A pump can be started and stopped multiple times.
 
 When a pump is first created it is _not_ started. You need to call the `start()` method to start it.
+<a href="mailto:julien@julienviet.com">Julien Viet</a>
