+++ src/main/java/io/vertx/core/Closeable.java
@@ -14,10 +14,7 @@
  * You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.core.impl;
-
-import io.vertx.core.AsyncResult;
-import io.vertx.core.Handler;
+package io.vertx.core;
 
 /**
  * @author <a href="http://tfox.org">Tim Fox</a>
