+++ src/main/java/io/vertx/core/eventbus/impl/clustered/ClusteredMessage.java
@@ -14,7 +14,7 @@
  * You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.core.eventbus.impl;
+package io.vertx.core.eventbus.impl.clustered;
 
 import io.netty.util.CharsetUtil;
 import io.vertx.core.AsyncResult;
@@ -24,12 +24,11 @@ import io.vertx.core.buffer.Buffer;
 import io.vertx.core.eventbus.DeliveryOptions;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.eventbus.MessageCodec;
-import io.vertx.core.eventbus.ReplyException;
-import io.vertx.core.eventbus.ReplyFailure;
+import io.vertx.core.eventbus.impl.CodecManager;
+import io.vertx.core.eventbus.impl.local.LocalMessage;
 import io.vertx.core.http.CaseInsensitiveHeaders;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
-import io.vertx.core.net.NetSocket;
 import io.vertx.core.net.impl.ServerID;
 
 import java.util.List;
@@ -38,77 +37,40 @@ import java.util.Map;
 /**
  * @author <a href="http://tfox.org">Tim Fox</a>
  */
-public class MessageImpl<U, V> implements Message<V> {
+public class ClusteredMessage<U, V> extends LocalMessage<U, V> {
 
-  private static final Logger log = LoggerFactory.getLogger(MessageImpl.class);
+  private static final Logger log = LoggerFactory.getLogger(ClusteredMessage.class);
 
   private static final byte WIRE_PROTOCOL_VERSION = 1;
 
-  private NetSocket socket;
-  private EventBusImpl bus;
   private ServerID sender;
-  private String address;
-  private String replyAddress;
-  private MultiMap headers;
-  private U sentBody;
-  private V receivedBody;
-  private MessageCodec<U, V> messageCodec;
-  private boolean send;
   private Buffer wireBuffer;
   private int bodyPos;
   private int headersPos;
+  private boolean fromWire;
 
-  public MessageImpl() {
+  public ClusteredMessage() {
   }
 
-  public MessageImpl(ServerID sender, String address, String replyAddress, MultiMap headers, U sentBody,
-                     MessageCodec<U, V> messageCodec, boolean send) {
+  public ClusteredMessage(ServerID sender, String address, String replyAddress, MultiMap headers, U sentBody,
+                          MessageCodec<U, V> messageCodec, boolean send) {
+    super(address, replyAddress, headers, sentBody, messageCodec, send);
     this.sender = sender;
-    this.address = address;
-    this.replyAddress = replyAddress;
-    this.headers = headers;
-    this.sentBody = sentBody;
-    this.messageCodec = messageCodec;
-    this.send = send;
   }
 
-  private MessageImpl(MessageImpl<U, V> other) {
-    this.socket = other.socket;
-    this.bus = other.bus;
+  protected ClusteredMessage(ClusteredMessage<U, V> other) {
+    super(other);
     this.sender = other.sender;
-    this.address = other.address;
-    this.replyAddress = other.replyAddress;
-    this.messageCodec = other.messageCodec;
-    if (other.headers != null) {
-      List<Map.Entry<String, String>> entries = other.headers.entries();
-      this.headers = new CaseInsensitiveHeaders();
-      for (Map.Entry<String, String> entry: entries) {
-        this.headers.add(entry.getKey(), entry.getValue());
-      }
-    }
-    if (other.sentBody != null) {
-      // This will only be true if the message has been sent locally
-      this.sentBody = other.sentBody;
-      this.receivedBody = messageCodec.transform(other.sentBody);
-    } else {
+    if (other.sentBody == null) {
       this.wireBuffer = other.wireBuffer;
       this.bodyPos = other.bodyPos;
       this.headersPos = other.headersPos;
     }
-    this.send = other.send;
+    this.fromWire = other.fromWire;
   }
 
-  NetSocket getSocket() {
-    return socket;
-  }
-
-  public MessageImpl<U, V> copyBeforeReceive() {
-    return new MessageImpl<>(this);
-  }
-
-  @Override
-  public String address() {
-    return address;
+  public ClusteredMessage<U, V> copyBeforeReceive() {
+    return new ClusteredMessage<>(this);
   }
 
   @Override
@@ -164,13 +126,10 @@ public class MessageImpl<U, V> implements Message<V> {
     encodeHeaders(buffer);
     writeBody(buffer);
     buffer.setInt(0, buffer.length() - 4);
-//    if (buffer.length()> length) {
-//      log.warn("Overshot length " + length + " actual " + buffer.length());
-//    }
     return buffer;
   }
 
-  public void readFromWire(NetSocket socket, Buffer buffer, Map<String, MessageCodec> codecMap, MessageCodec[] systemCodecs) {
+  public void readFromWire(Buffer buffer, CodecManager codecManager) {
     int pos = 0;
     // Overall Length already read when passed in here
     byte protocolVersion = buffer.getByte(pos);
@@ -187,13 +146,13 @@ public class MessageImpl<U, V> implements Message<V> {
       pos += 4;
       byte[] bytes = buffer.getBytes(pos, pos + length);
       String codecName = new String(bytes, CharsetUtil.UTF_8);
-      messageCodec = codecMap.get(codecName);
+      messageCodec = codecManager.getCodec(codecName);
       if (messageCodec == null) {
         throw new IllegalStateException("No message codec registered with name " + codecName);
       }
       pos += length;
     } else {
-      messageCodec = systemCodecs[systemCodecCode];
+      messageCodec = codecManager.systemCodecs()[systemCodecCode];
     }
     byte bsend = buffer.getByte(pos);
     send = bsend == 0;
@@ -223,7 +182,7 @@ public class MessageImpl<U, V> implements Message<V> {
     bodyPos = pos;
     sender = new ServerID(senderPort, senderHost);
     wireBuffer = buffer;
-    this.socket = socket;
+    fromWire = true;
   }
 
   private void decodeBody() {
@@ -283,56 +242,13 @@ public class MessageImpl<U, V> implements Message<V> {
   }
 
   @Override
-  public void fail(int failureCode, String message) {
-    if (replyAddress != null) {
-      sendReply(bus.createMessage(true, replyAddress, null,
-          new ReplyException(ReplyFailure.RECIPIENT_FAILURE, failureCode, message), null), null, null);
+  protected <R> void sendReply(LocalMessage msg, DeliveryOptions options, Handler<AsyncResult<Message<R>>> replyHandler) {
+    if (bus != null & replyAddress() != null) {
+      ((ClusteredEventBus)bus).sendReply(sender, msg, options, replyHandler);
     }
   }
 
-  @Override
-  public void reply(Object message) {
-    reply(message, new DeliveryOptions(), null);
+  public boolean isFromWire() {
+    return fromWire;
   }
-
-  @Override
-  public <R> void reply(Object message, Handler<AsyncResult<Message<R>>> replyHandler) {
-    reply(message, new DeliveryOptions(), replyHandler);
-  }
-
-  @Override
-  public void reply(Object message, DeliveryOptions options) {
-    reply(message, options, null);
-  }
-
-  @Override
-  public <R> void reply(Object message, DeliveryOptions options, Handler<AsyncResult<Message<R>>> replyHandler) {
-    if (replyAddress != null) {
-      sendReply(bus.createMessage(true, replyAddress, options.getHeaders(), message, options.getCodecName()), options, replyHandler);
-    }
-  }
-
-  protected void setReplyAddress(String replyAddress) {
-    this.replyAddress = replyAddress;
-  }
-
-  protected boolean send() {
-    return send;
-  }
-
-  protected void setBus(EventBusImpl eventBus) {
-    this.bus = eventBus;
-  }
-
-  protected MessageCodec codec() {
-    return messageCodec;
-  }
-
-  private <R> void sendReply(MessageImpl msg, DeliveryOptions options, Handler<AsyncResult<Message<R>>> replyHandler) {
-    if (bus != null && replyAddress != null) {
-      bus.sendReply(sender, msg, options, replyHandler);
-    }
-  }
-
-
 }
