+++ src/main/java/io/vertx/core/http/impl/HttpServerImpl.java
@@ -38,7 +38,7 @@ import io.vertx.core.http.*;
 import io.vertx.core.http.impl.cgbystrom.FlashPolicyHandler;
 import io.vertx.core.http.impl.ws.WebSocketFrameImpl;
 import io.vertx.core.http.impl.ws.WebSocketFrameInternal;
-import io.vertx.core.impl.Closeable;
+import io.vertx.core.Closeable;
 import io.vertx.core.impl.ContextImpl;
 import io.vertx.core.impl.VertxInternal;
 import io.vertx.core.logging.Logger;
@@ -255,17 +255,7 @@ public class HttpServerImpl implements HttpServer, Closeable, MetricsProvider {
             res = Future.failedFuture(future.cause());
             listening = false;
           }
-          // FIXME - workaround for https://github.com/netty/netty/issues/2586
-          // If listen already succeeded on a different event loop, and then addListener is called again
-          // on the completed future from a different event loop then the handler will be called on the original
-          // event loop not on the when that called addListener.
-          // To reproduce set the boolean parameter on execute (below) to true.
-          // Then run Httptest.testTwoServersSameAddressDifferentContext()
-          try {
-            listenContext.runOnContext((v) -> listenHandler.handle(res));
-          } catch (Exception e) {
-            e.printStackTrace();
-          }
+          listenContext.runOnContext((v) -> listenHandler.handle(res));
         } else if (!future.isSuccess()) {
           listening  = false;
           // No handler - log so user can see failure
