+++ src/main/java/io/vertx/core/net/impl/NetServerImpl.java
@@ -37,7 +37,7 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.AsyncResultHandler;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
-import io.vertx.core.impl.Closeable;
+import io.vertx.core.Closeable;
 import io.vertx.core.impl.ContextImpl;
 import io.vertx.core.impl.VertxInternal;
 import io.vertx.core.logging.Logger;
