+++ src/main/resources/META-INF/services/io.vertx.core.spi.EventBusFactory
@@ -1,0 +1,1 @@
+io.vertx.core.eventbus.impl.EventBusFactoryImpl
