+++ src/test/asciidoc/dataobjects.adoc
@@ -1,19 +1,5 @@
 = Cheatsheets
 
-[[NoConverterDataObject]]
-== NoConverterDataObject
-
-++++
- @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
-++++
-'''
-
-[cols=">25%,^25%,50%"]
-[frame="topbot"]
-|===
-^|Name | Type ^| Description
-|===
-
 [[ChildInheritingDataObject]]
 == ChildInheritingDataObject
 
@@ -30,8 +16,8 @@
 |[[parentProperty]]`parentProperty`|`String`|-
 |===
 
-[[ParentDataObject]]
-== ParentDataObject
+[[NoConverterDataObject]]
+== NoConverterDataObject
 
 ++++
  @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
@@ -42,7 +28,6 @@
 [frame="topbot"]
 |===
 ^|Name | Type ^| Description
-|[[parentProperty]]`parentProperty`|`String`|-
 |===
 
 [[TestDataObject]]
@@ -127,6 +112,21 @@
 |[[stringValue]]`stringValue`|`String`|-
 |[[stringValueMap]]`stringValueMap`|`String`|-
 |[[stringValues]]`stringValues`|`Array of String`|-
+|===
+
+[[ParentDataObject]]
+== ParentDataObject
+
+++++
+ @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
+++++
+'''
+
+[cols=">25%,^25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[parentProperty]]`parentProperty`|`String`|-
 |===
 
 [[AggregatedDataObject]]
